Move assessment overview card to separate file
Also refactor to React.FC.
